requirements-ocr.txt
easyocr>=1.7.1
pillow>=10.0.0
opencv-python>=4.8.0.76
pytesseract>=0.3.10
